remove temp directory if cloned successfully

+++ install.sh
@@ -2,8 +2,8 @@
 
 ( cd /tmp/ && \
 git clone https://github.com/Logicer16/pam-watchid.git && \
-cd "$(basename "$_" .git)" && \
-make $1 && \
+cd "$(basename "$_" .git)" && CLONE_SUCCESS="true" && \ 
+make $1
 # cleanup
-TMP=`pwd -P` && cd "`dirname $TMP`" && rm -rf "./`basename $TMP`" && unset TMP
+[[ "$CLONE_SUCCESS" == "true" ]] && TMP=`pwd -P` && cd "`dirname $TMP`" && rm -rf "./`basename $TMP`" && unset TMP
 )
